:recycle: theme: refactor color to be under theme
so that we can add icon theme
